memory: fixed dogecoin_btree_tdestroy
still need to look into dogecoin_wallet_free as to what was being
passed that caused the error in the first place but for now the
dogecoin_btree_tdestroy function, which is used by many functions
in our library, has as many guards as i can think of to prevent it
recursively calling itself when containing a 'nil' value on root
node->right. first we check if root node exists in memory, if so
then proceed to evaluate if freekey is populated and if true we
jump over everything to free the key and root node itself, otherwise
we still have work to do. next step is to determine if root node
left and right are 0 and if true return, if not then if root node
left exists repeat first step and recursively call on self and jump
past evaluation section. finally if r->right equals 'nil' we perform
one last destroy and proceed to recursively call ourself until we hit
if (r == 0) return and exit successfully destroying our red black tree.