security: cwe-14
sensitive memory is cleared according to the source code, but compiler optimizations leave the memory untouched when it is not read from again, aka dead store removal.

there are about 3 remaining memset's on l207 in base58.c, l166 in protocol.c and l28 in ecc_tests.c that still need to be addressed.